были удалены ненужные классы, добавлен WatchingTower, добавлена Alive(проверка на то, жив юнит или нет), начата работа над оружием, рассортированно по папкам
