--- UnitsDrafts/BuildingFactory.cs
@@ -1,10 +1,0 @@
-﻿namespace UnitsDrafts
-{
-    internal static class BuildingFactory
-    {
-        public static void CreateBarracs()
-        {
-
-        }
-    }
-}
