+++ UnitsDrafts/buildings/Barracs.cs
@@ -16,7 +16,7 @@ namespace UnitsDrafts
             Statistics.FootmanCount++;
             return new Footman();
         }
-        public Archer CrateArcher()
+        public Archer CreateArcher()
         {
             Statistics.ArcherCount++;
             return new Archer();
